this should fix the problems with the _make_error method.  Basically, it was generating an error and then the Parser would die.  If we were in the comp call process we'd a get a stack trace stuck on the end which would subsequently be eval'd and everything would blow up.
The fix (otherwise known as gross hack) is make the error string
"MASON: $error :NOSAM" so that it is enclosed on both sides by an
identifiable string.  I wrote a test for this (that uses the new test
suite) and was able to make it pass using this new code.


git-svn-id: https://svn.urth.org/svn/Mason/branches/mason-0-8-branch@576 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9