Eliminated several unneccesary eval("...") statements in capture_debug_state().  They were being used to run $r->$method, where $method isn't known until runtime.  However, $r->$method() is perfectly valid Perl syntax for this purpose.
The first eval() was also being used to report errors, so I left a
block-eval{} there.  The other places were not checking $@, so now
instead of failing silently they will issue an error if a particular
method isn't found.

This should result in a slight speedup when running under the
debugger, because eval("...") is slightly expensive.


git-svn-id: https://svn.urth.org/svn/Mason/trunk@613 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9