The source_callback routine can now return either a reference to the component source, or the source itself.
Added the comp_source_ref() method, which returns a reference to the
source rather than the source string itself.

Together, these changes allow better memory efficiency (and presumably
better speed, though that won't likely be noticable).

comp_source() is still the same as it always was, i.e. it returns the
source as a string.  It's now implemented as a wrapper around
comp_source_ref(), though.

These various ways of doing things are to preserve backward
compatibility.


git-svn-id: https://svn.urth.org/svn/Mason/trunk@2820 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9