compile() now passes through all arguments it receives to its superclass method (except comp_class - what's so special about that one?).
compiled_component() now accepts a 'fh' argument.  If present, it
represents a filehandle to which the compiled output should be
written.  It uses a new utility method _output_chunk() to manage this,
so that the output can either go to a scalar or a filehandle.  (funny,
it seems like a Buffer object could do this too...)

compile_to_file() can now emit its chunks directly to a file, by
passing an 'fh' parameter to compile().


git-svn-id: https://svn.urth.org/svn/Mason/trunk@2825 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9