Reverted back to Ken's version with an extra comment.
Detailed information from Ken:
"It uses foreach because join('', ...) uses more memory.  I turned it
into foreach when I was working on reducing memory requirements for the
Really Large Components, but it affects all components of any size.
join('', ...) uses more memory because it must allocate a temporary SV
to store the result of the join, then append that SV's contents to the
buffer.  The foreach version doesn't need any temporary internal
variables."


git-svn-id: https://svn.urth.org/svn/Mason/trunk@3067 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9