Take out the use of alarm() and SIG{ALRM} to trap rare infinite loops inside $self->_do_or_eval, because:
1) It interferes with surrounding Mason environments that use
alarm() for their own purposes.
2) The infinite loop can no longer be reproduced by any members of the
development team, meaning that it is even more rare than originally
thought (or perhaps depends on a certain combination of module
versions, lunar calendar, etc.)
3) The same "bad" code has been shown to seg fault rather than
infinite loop in certain environments; the alarm() will not help this
situation, and any tests we try to write for our alarm() trap
will seg fault themselves in these environments.
4) There are no doubt other Perl bugs lying out there now and in the
future that will cause infinite loops and other things, both in
compilation of components and in other parts of Mason.  There's no way
for us to anticipate all of these with alarm() traps, NOR is it our
responsibility to do so.


git-svn-id: https://svn.urth.org/svn/Mason/trunk@3126 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9