Use $r->notes() to track whether or not we've sent the headers in ApacheHandler. This replaces the hacky of looking at and messing with $r->content_type, which doesn't play well with apps like Bricolage, which want to set the content_type before passing control to Mason.
There was also an issue with sending headers twice in Mason 1.36, and
this is fixed as well.

Added tests for all of this, to make sure that we preserve a
Content-type set before Mason is called, and that only one set of
headers is sent.

Also enabled Apache::Filter test for maintainers running 08-ah.t,
since this was never getting run for anyone previously.


git-svn-id: https://svn.urth.org/svn/Mason/trunk@3841 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9