A huge refactoring of the incredibly nasty Apache live tests.
This moves the helper code into a proper library.

One big improvement is that test counting is done in the helper
library, so there's no need to manually recalculate it every time we
want to add a test.

Additionally, the sets of tests to be run are grouped a little more
logically, without lots of conditionals.

The actual test execution code was ridiculously full of duplication,
which is now gone.

Each different Apache config now gets its own .t file.

Overall, this is much cleaner, though it's still not all that
nice. But is should make adding new tests a lot easier, which was the
biggest problem with the existing code in my opinion.

A few things still todo:

- There is a test that fails for some configurations, for explicitly
  calling $r->send_http_header() in a component. This is a known bug,
  which I plan to fix.

- The live-cgi tests need to be similarly refactored.


git-svn-id: https://svn.urth.org/svn/Mason/trunk@3852 fa5c2c93-f9fe-0310-b630-c7cdf19d91a9